refactor: simplify test naming and category table
