+++ asyncband/src/waitgroup/tests.rs
@@ -15,17 +15,19 @@
 // specific language governing permissions and limitations
 // under the License.
 
-use super::Semaphore;
+use std::sync::Arc;
+
+use super::WaitGroup;
+use crate::internal::CountdownState;
+
+// This test stays next to the implementation because it inspects private state.
 
 #[test]
-fn fulfilled_forget_exact_debt_reclaims_its_waiter_node() {
-    let semaphore = Semaphore::new(0);
+#[should_panic(expected = "WaitGroup counter overflow")]
+fn clone_panics_on_counter_overflow() {
+    let wg = WaitGroup {
+        state: Arc::new(CountdownState::new(u32::MAX)),
+    };
 
-    for _ in 0..3 {
-        semaphore.forget_exact(1);
-        assert_eq!(semaphore.s.num_waiter_nodes(), 1);
-
-        semaphore.release(1);
-        assert_eq!(semaphore.s.num_waiter_nodes(), 0);
-    }
+    let _ = wg.clone();
 }
